Separate the calculation of rotation derivative into a new header
