+++ lammps/lib/colvars/Makefile.deps
@@ -3,7 +3,7 @@ $(COLVARS_OBJ_DIR)colvaratoms.o: colvaratoms.cpp colvarmodule.h \
  colvars_version.h colvarproxy.h colvartypes.h colvarvalue.h \
  colvarproxy_io.h colvarproxy_system.h colvarproxy_tcl.h \
  colvarproxy_volmaps.h colvarparse.h colvarparams.h colvaratoms.h \
- colvardeps.h
+ colvardeps.h colvar_rotation_derivative.h
 $(COLVARS_OBJ_DIR)colvarbias_abf.o: colvarbias_abf.cpp colvarmodule.h \
  colvars_version.h colvar.h colvarvalue.h colvartypes.h colvarparse.h \
  colvarparams.h colvardeps.h colvarbias_abf.h colvarproxy.h \
@@ -71,7 +71,7 @@ $(COLVARS_OBJ_DIR)colvarcomp_distances.o: colvarcomp_distances.cpp \
  colvarparse.h colvarparams.h colvar.h colvardeps.h colvarcomp.h \
  colvaratoms.h colvarproxy.h colvarproxy_io.h colvarproxy_system.h \
  colvarproxy_tcl.h colvarproxy_volmaps.h colvar_arithmeticpath.h \
- colvar_geometricpath.h
+ colvar_geometricpath.h colvar_rotation_derivative.h
 $(COLVARS_OBJ_DIR)colvarcomp_gpath.o: colvarcomp_gpath.cpp colvarmodule.h \
  colvars_version.h colvarvalue.h colvartypes.h colvarparse.h \
  colvarparams.h colvar.h colvardeps.h colvarcomp.h colvaratoms.h \
@@ -100,7 +100,7 @@ $(COLVARS_OBJ_DIR)colvarcomp_rotations.o: colvarcomp_rotations.cpp \
  colvarparse.h colvarparams.h colvar.h colvardeps.h colvarcomp.h \
  colvaratoms.h colvarproxy.h colvarproxy_io.h colvarproxy_system.h \
  colvarproxy_tcl.h colvarproxy_volmaps.h colvar_arithmeticpath.h \
- colvar_geometricpath.h
+ colvar_geometricpath.h colvar_rotation_derivative.h
 $(COLVARS_OBJ_DIR)colvarcomp_volmaps.o: colvarcomp_volmaps.cpp \
  colvarmodule.h colvars_version.h colvarvalue.h colvartypes.h \
  colvarparse.h colvarparams.h colvar.h colvardeps.h colvarcomp.h \
@@ -191,7 +191,7 @@ $(COLVARS_OBJ_DIR)colvarscript_commands_colvar.o: \
  colvarscript_commands_bias.h
 $(COLVARS_OBJ_DIR)colvartypes.o: colvartypes.cpp colvarmodule.h \
  colvars_version.h colvartypes.h colvarparse.h colvarvalue.h \
- colvarparams.h ../../src/math_eigen_impl.h
+ colvarparams.h colvar_rotation_derivative.h ../../src/math_eigen_impl.h
 $(COLVARS_OBJ_DIR)colvarvalue.o: colvarvalue.cpp colvarmodule.h \
  colvars_version.h colvarvalue.h colvartypes.h
 $(COLVARS_OBJ_DIR)colvar_neuralnetworkcompute.o: \
