+++ namd/colvars/Make.depends
@@ -39,7 +39,8 @@ obj/colvaratoms.o: \
 	colvars/src/colvarparse.h \
 	colvars/src/colvarparams.h \
 	colvars/src/colvaratoms.h \
-	colvars/src/colvardeps.h
+	colvars/src/colvardeps.h \
+        colvars/src/colvar_rotation_derivative.h
 	$(CXX) $(COLVARSCXXFLAGS) $(COPTO)obj/colvaratoms.o $(COPTC) colvars/src/colvaratoms.cpp
 obj/colvarbias.o: \
 	obj/.exists \
@@ -299,7 +300,8 @@ obj/colvarcomp_distances.o: \
 	colvars/src/colvarproxy_tcl.h \
 	colvars/src/colvarproxy_volmaps.h \
 	colvars/src/colvar_arithmeticpath.h \
-	colvars/src/colvar_geometricpath.h
+	colvars/src/colvar_geometricpath.h \
+        colvars/src/colvar_rotation_derivative.h
 	$(CXX) $(COLVARSCXXFLAGS) $(COPTO)obj/colvarcomp_distances.o $(COPTC) colvars/src/colvarcomp_distances.cpp
 obj/colvarcomp_gpath.o: \
 	obj/.exists \
@@ -362,7 +364,8 @@ obj/colvarcomp_rotations.o: \
 	colvars/src/colvarproxy_tcl.h \
 	colvars/src/colvarproxy_volmaps.h \
 	colvars/src/colvar_arithmeticpath.h \
-	colvars/src/colvar_geometricpath.h
+	colvars/src/colvar_geometricpath.h \
+        colvars/src/colvar_rotation_derivative.h
 	$(CXX) $(COLVARSCXXFLAGS) $(COPTO)obj/colvarcomp_rotations.o $(COPTC) colvars/src/colvarcomp_rotations.cpp
 obj/colvarcomp_volmaps.o: \
 	obj/.exists \
@@ -711,7 +714,8 @@ obj/colvartypes.o: \
 	colvars/src/colvarparse.h \
 	colvars/src/colvarvalue.h \
 	colvars/src/colvarparams.h \
-	colvars/src/nr_jacobi.h
+	colvars/src/nr_jacobi.h \
+        colvars/src/colvar_rotation_derivative.h
 	$(CXX) $(COLVARSCXXFLAGS) $(COPTO)obj/colvartypes.o $(COPTC) colvars/src/colvartypes.cpp
 obj/colvarvalue.o: \
 	obj/.exists \
