+++ src/colvaratoms.h
@@ -15,6 +15,9 @@
 #include "colvarparse.h"
 #include "colvardeps.h"
 
+template <typename T1, typename T2>
+struct rotation_derivative;
+
 
 /// \brief Stores numeric id, mass and all mutable data for an atom,
 /// mostly used by a \link colvar::cvc \endlink
@@ -331,7 +334,7 @@ public:
   cvm::rotation rot;
 
   /// Rotation derivative;
-  cvm::rotation::derivative<cvm::atom, cvm::atom_pos>* rot_deriv;
+  rotation_derivative<cvm::atom, cvm::atom_pos>* rot_deriv;
 
   /// \brief Indicates that the user has explicitly set centerToReference or
   /// rotateReference, and the corresponding reference:
@@ -496,103 +499,9 @@ public:
 
   /// \brief Calculate the derivatives of the fitting transformation
   void calc_fit_gradients();
-  /*! @brief  Actual implementation of `calc_fit_gradients`. The template is
-   *          used to avoid branching inside the loops in case that the CPU
-   *          branch prediction is broken (or further migration to GPU code).
-   *  @tparam B_ag_center Centered the reference to origin? This should follow
-   *          the value of `is_enabled(f_ag_center)`.
-   *  @tparam B_ag_rotate Calculate the optimal rotation? This should follow
-   *          the value of `is_enabled(f_ag_rotate)`.
-   */
+
   template <bool B_ag_center, bool B_ag_rotate>
-  void calc_fit_gradients_impl() {
-    cvm::atom_group *group_for_fit = fitting_group ? fitting_group : this;
-    // the center of geometry contribution to the gradients
-    cvm::rvector atom_grad;
-    // the rotation matrix contribution to the gradients
-    cvm::rotation const rot_inv = rot.inverse();
-    // temporary variables for computing and summing derivatives
-    cvm::real sum_dxdq[4] = {0, 0, 0, 0};
-    cvm::vector1d<cvm::rvector> dq0_1(4);
-    // loop 1: iterate over the current atom group
-    for (size_t i = 0; i < this->size(); i++) {
-      cvm::atom_pos pos_orig;
-      if (B_ag_center) {
-        atom_grad += atoms[i].grad;
-        if (B_ag_rotate) pos_orig = rot_inv.rotate(atoms[i].pos - ref_pos_cog);
-      } else {
-        if (B_ag_rotate) pos_orig = atoms[i].pos;
-      }
-      if (B_ag_rotate) {
-        // calculate \partial(R(q) \vec{x}_i)/\partial q) \cdot \partial\xi/\partial\vec{x}_i
-        cvm::quaternion const dxdq =
-          rot.q.position_derivative_inner(pos_orig, atoms[i].grad);
-        sum_dxdq[0] += dxdq[0];
-        sum_dxdq[1] += dxdq[1];
-        sum_dxdq[2] += dxdq[2];
-        sum_dxdq[3] += dxdq[3];
-      }
-    }
-    if (B_ag_center) {
-      if (B_ag_rotate) atom_grad = (rot.inverse()).rotate(atom_grad);
-      atom_grad *= (-1.0)/(cvm::real(group_for_fit->size()));
-    }
-    // loop 2: iterate over the fitting group
-    if (B_ag_rotate) rot_deriv->prepare_derivative(false, true);
-    for (size_t j = 0; j < group_for_fit->size(); j++) {
-      if (B_ag_center) {
-        group_for_fit->fit_gradients[j] = atom_grad;
-      }
-      if (B_ag_rotate) {
-        rot_deriv->calc_derivative_to_group1(j, NULL, &dq0_1);
-        // multiply by {\partial q}/\partial\vec{x}_j and add it to the fit gradients
-        group_for_fit->fit_gradients[j] += sum_dxdq[0] * dq0_1[0] +
-                                           sum_dxdq[1] * dq0_1[1] +
-                                           sum_dxdq[2] * dq0_1[2] +
-                                           sum_dxdq[3] * dq0_1[3];
-      }
-    }
-    /* The above code is an optimized version of the following loops:
-    if (is_enabled(f_ag_center)) {
-      // add the center of geometry contribution to the gradients
-      cvm::rvector atom_grad;
-
-      for (size_t i = 0; i < this->size(); i++) {
-        atom_grad += atoms[i].grad;
-      }
-      if (is_enabled(f_ag_rotate)) atom_grad = (rot.inverse()).rotate(atom_grad);
-      atom_grad *= (-1.0)/(cvm::real(group_for_fit->size()));
-
-      for (size_t j = 0; j < group_for_fit->size(); j++) {
-        group_for_fit->fit_gradients[j] = atom_grad;
-      }
-    }
-
-    if (is_enabled(f_ag_rotate)) {
-
-      // add the rotation matrix contribution to the gradients
-      cvm::rotation const rot_inv = rot.inverse();
-
-      for (size_t i = 0; i < this->size(); i++) {
-
-        // compute centered, unrotated position
-        cvm::atom_pos const pos_orig =
-          rot_inv.rotate((is_enabled(f_ag_center) ? (atoms[i].pos - ref_pos_cog) : (atoms[i].pos)));
-
-        // calculate \partial(R(q) \vec{x}_i)/\partial q) \cdot \partial\xi/\partial\vec{x}_i
-        cvm::quaternion const dxdq =
-          rot.q.position_derivative_inner(pos_orig, atoms[i].grad);
-
-        for (size_t j = 0; j < group_for_fit->size(); j++) {
-          // multiply by {\partial q}/\partial\vec{x}_j and add it to the fit gradients
-          for (size_t iq = 0; iq < 4; iq++) {
-            group_for_fit->fit_gradients[j] += dxdq[iq] * rot.dQ0_1[j][iq];
-          }
-        }
-      }
-    }
-   */
-  }
+  friend void calc_fit_gradients_impl(colvarmodule::atom_group* ag);
 
   /// \brief Derivatives of the fitting transformation
   std::vector<cvm::atom_pos> fit_gradients;
