+++ src/colvarcomp_rotations.cpp
@@ -12,7 +12,7 @@
 #include "colvarparse.h"
 #include "colvar.h"
 #include "colvarcomp.h"
-
+#include "colvar_rotation_derivative.h"
 
 
 colvar::orientation::orientation(std::string const &conf)
@@ -127,11 +127,11 @@ void colvar::orientation::apply_force(colvarvalue const &force)
   cvm::quaternion const &FQ = force.quaternion_value;
 
   if (!atoms->noforce) {
-    cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+    rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
     rot_deriv.prepare_derivative(false, true);
     cvm::vector1d<cvm::rvector> dq0_2;
     for (size_t ia = 0; ia < atoms->size(); ia++) {
-      rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+      rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
       for (size_t i = 0; i < 4; i++) {
         (*atoms)[ia].apply_force(FQ[i] * dq0_2[i]);
       }
@@ -200,11 +200,11 @@ void colvar::orientation_angle::calc_gradients()
       ((180.0 / PI) * (-2.0) / cvm::sqrt(1.0 - ((rot.q).q0 * (rot.q).q0))) :
       0.0 );
 
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     (*atoms)[ia].grad = (dxdq0 * dq0_2[0]);
   }
 }
@@ -252,11 +252,11 @@ void colvar::orientation_proj::calc_value()
 void colvar::orientation_proj::calc_gradients()
 {
   cvm::real const dxdq0 = 2.0 * 2.0 * (rot.q).q0;
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     (*atoms)[ia].grad = (dxdq0 * dq0_2[0]);
   }
 }
@@ -318,12 +318,12 @@ void colvar::tilt::calc_gradients()
 {
   cvm::quaternion const dxdq = rot.dcos_theta_dq(axis);
 
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
     (*atoms)[ia].grad = cvm::rvector(0.0, 0.0, 0.0);
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     for (size_t iq = 0; iq < 4; iq++) {
       (*atoms)[ia].grad += (dxdq[iq] * dq0_2[iq]);
     }
@@ -399,12 +399,12 @@ void colvar::spin_angle::calc_gradients()
 {
   cvm::quaternion const dxdq = rot.dspin_angle_dq(axis);
 
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
     (*atoms)[ia].grad = cvm::rvector(0.0, 0.0, 0.0);
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     for (size_t iq = 0; iq < 4; iq++) {
       (*atoms)[ia].grad += (dxdq[iq] * dq0_2[iq]);
     }
@@ -520,11 +520,11 @@ void colvar::euler_phi::calc_gradients()
   const cvm::real dxdq1 = (180.0/PI) * (2 * q0 * (-2 * q1 * q1 - 2 * q2 * q2 + 1) - 4 * q1 * (-2 * q0 * q1 - 2 * q2 * q3)) / denominator;
   const cvm::real dxdq2 = (180.0/PI) * (-4 * q2 * (-2 * q0 * q1 - 2 * q2 * q3) + 2 * q3 * (-2 * q1 * q1 - 2 * q2 * q2 + 1)) / denominator;
   const cvm::real dxdq3 = (180.0/PI) * 2 * q2 * (-2 * q1 * q1 - 2 * q2 * q2 + 1) / denominator;
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     (*atoms)[ia].grad = (dxdq0 * dq0_2[0]) +
                         (dxdq1 * dq0_2[1]) +
                         (dxdq2 * dq0_2[2]) +
@@ -640,11 +640,11 @@ void colvar::euler_psi::calc_gradients()
   const cvm::real dxdq1 = (180.0/PI) * 2 * q2 * (-2 * q2 * q2 - 2 * q3 * q3 + 1) / denominator;
   const cvm::real dxdq2 = (180.0/PI) * (2 * q1 * (-2 * q2 * q2 - 2 * q3 * q3 + 1) - 4 * q2 * (-2 * q0 * q3 - 2 * q1 * q2)) / denominator;
   const cvm::real dxdq3 = (180.0/PI) * (2 * q0 * (-2 * q2 * q2 - 2 * q3 * q3 + 1) - 4 * q3 * (-2 * q0 * q3 - 2 * q1 * q2)) / denominator;
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     (*atoms)[ia].grad = (dxdq0 * dq0_2[0]) +
                         (dxdq1 * dq0_2[1]) +
                         (dxdq2 * dq0_2[2]) +
@@ -758,11 +758,11 @@ void colvar::euler_theta::calc_gradients()
   const cvm::real dxdq1 = (180.0/PI) * -2 * q3 / denominator;
   const cvm::real dxdq2 = (180.0/PI) * 2 * q0 / denominator;
   const cvm::real dxdq3 = (180.0/PI) * -2 * q1 / denominator;
-  cvm::rotation::derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
+  rotation_derivative<cvm::atom_pos, cvm::atom_pos> rot_deriv(rot, ref_pos, shifted_pos);
   rot_deriv.prepare_derivative(false, true);
   cvm::vector1d<cvm::rvector> dq0_2;
   for (size_t ia = 0; ia < atoms->size(); ia++) {
-    rot_deriv.calc_derivative_to_group2(ia, NULL, &dq0_2);
+    rot_deriv.calc_derivative_to_group2(ia, nullptr, &dq0_2);
     (*atoms)[ia].grad = (dxdq0 * dq0_2[0]) +
                         (dxdq1 * dq0_2[1]) +
                         (dxdq2 * dq0_2[2]) +
